[IMP] website: add specific object for Optimize SEO
Before this commit, the object used by Optimize SEO 'wizard' was always the
main object, without any possibility to change that (except by changing the
main object but that would cause a lot of other issue, eg edit in backend).

This commit add the possibility to explicitly set a seo object on the html data
attribute.
If set, this one will be used instead of the main object.

This is useful for event.event where all the event pages (with website menu
enabled) would share the same SEO since the main object would be the
event.event.
With this commit, the seo object is now the website.page, thus giving the
possibility to have a separate SEO for every event pages.

(Required for Odoo.com website for OXP)

closes odoo/odoo#33194

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>