[FIX] sale_timesheet: default filters to setup services
In settings, we added the possibility to edit service products per type
(time-based, fixed, or milestone). An action was trigger and a default filter
was applied. This filter was removed with 3ef01cf0eea154f8cba6988b8bbc38950066d1dd, making the 3 settings
links displaying all products.
This commit restore the filters. In order to keep the UI silmplification,
those 3 filters are only displayed (in search view) for this specific flow.
So, we keep the general search view simple.

opw-1987352