[FIX] mail: remove readonly on recommended next activity
Fine-tuning of 595c48c056950cbd17f2d81f8d51e9c459b33384.
It does not make sense to be able to write on mail.activity.type
using the related fields defined on mail.activity.

But this field is not a related one, and it is necessary to be writable to make
the 'recommended next activity' feature work at all.

opw 1974236

closes odoo/odoo#33325

Signed-off-by: Nans Lefebvre (len) <len@odoo.com>