[FIX] hr_holidays: request leave with leave type in hours
When a leave type is in hours, before this commit:
- in the leaves request list the request duration is shown in days.
- the name of the leave shown duration in days.
- the rounding after the conversion from hours to days, generates some
misunderstandings. If you have a day of 7.5 hours, and you take a leave
of 2 hours, it would be round it to 0.25 days, and not to 0.27 days as
it should be. This occurs because we round to the 1/16 of the day.

Now:
- in the leave request list, the request is shown either with days or
hours depending on the leave type.
- the name of the leave shown duration either with days or hours
depending on the leave type.
- the leaves are shown in the correct unit, there is not more
misunderstandings occurring because of the rounding.

related to commit : 08066706e4bd3ac83137757f5694450e1a5730f3 and 48577fa610264e7ad8de38ba9d30fc0fb45a0785

opw-1974691

closes #29804
closes #33290

closes odoo/odoo#33369

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>