[IMP] crm: allow onchange of customer from lead quick create
Before commit

When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" doesn't
appear in the "Followup" tab on the lead so we have to reassign a partner
second time on the opportunity.

After commit

When we create an opportunity manually from the Kanban view and associate
a customer to the opportunity, the "Initial Contact Information" will appear
in the "Followup" tab on lead.

This commit adds missing fields as invisible in the quick create view so
that the onchange updates them. As original author say, "to make onchange
works fabulously".

Task ID 1974504